+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 - Greasemonkey scripts that broke when the header was directly next to the script
 - Greasemonkey scripts not working if multiple match the same domain by using an IIFE
+- Setting "mouse" not working at all in 11.0.0 due to incorrect setting type
 - Downloads not kept on restart since V10
 
 ### Versions